:sparkles: feat: Add websocket layer to push notifications to front-end

+++ services/order_service/app/core/db/models/order_item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import ForeignKey, Integer
+from sqlalchemy import Float, ForeignKey, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 from services.order_service.app.core.db.base import Base
 
@@ -9,5 +9,5 @@ class OrderItem(Base):
     order_id: Mapped[int] = mapped_column(ForeignKey("orders.id"), primary_key=True, nullable=False)
     dish_id: Mapped[int] = mapped_column(ForeignKey("dishes.id"), primary_key=True, nullable=False)
     quantity: Mapped[int] = mapped_column(Integer, nullable=False)
-    total: Mapped[int] = mapped_column(Integer, nullable=False)
+    total: Mapped[float] = mapped_column(Float, nullable=False)
